+++ realm/realm-library/src/androidTestObjectServer/kotlin/io/realm/EmailPasswordAuthTests.kt
@@ -31,7 +31,6 @@ import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.lang.IllegalStateException
 import kotlin.test.assertFailsWith
 
 @RunWith(AndroidJUnit4::class)
@@ -58,6 +57,7 @@ class EmailPasswordAuthTests {
         RESEND_CONFIRMATION_EMAIL,
         SEND_RESET_PASSWORD_EMAIL,
         CALL_RESET_PASSWORD_FUNCTION,
+        RETRY_CUSTOM_CONFIRMATION,
         RESET_PASSWORD
     }
 
@@ -221,7 +221,7 @@ class EmailPasswordAuthTests {
                 val provider = app.emailPassword
                 provider.registerUser(email, "123456")
                 provider.resendConfirmationEmailAsync(email) { result ->
-                    when(result.isSuccess) {
+                    when (result.isSuccess) {
                         true -> looperThread.testComplete()
                         false -> fail(result.error.toString())
                     }
@@ -280,6 +280,114 @@ class EmailPasswordAuthTests {
     }
 
     @Test
+    fun retryCustomConfirmation() {
+        val email = "test_realm_tests_do_autoverify@10gen.com"
+        admin.setAutomaticConfirmation(false)
+        try {
+            val provider = app.emailPassword
+            provider.registerUser(email, "123456")
+            admin.setCustomConfirmation(true)
+
+            provider.retryCustomConfirmation(email)
+        } finally {
+            admin.setCustomConfirmation(false)
+        }
+    }
+
+    @Test
+    fun retryCustomConfirmation_failConfirmation() {
+        // Only emails containing realm_tests_do_autoverify will be confirmed
+        val email = "test@10gen.com"
+        admin.setAutomaticConfirmation(false)
+        try {
+            val provider = app.emailPassword
+            provider.registerUser(email, "123456")
+            admin.setCustomConfirmation(true)
+
+            val exception = assertFailsWith<AppException> {
+                provider.retryCustomConfirmation(email)
+            }
+
+            assertEquals("failed to confirm user test@10gen.com", exception.errorMessage)
+
+        } finally {
+            admin.setCustomConfirmation(false)
+        }
+    }
+
+    @Test
+    fun retryCustomConfirmationAsync() {
+        val email = "test_realm_tests_do_autoverify@10gen.com"
+        admin.setAutomaticConfirmation(false)
+        try {
+            looperThread.runBlocking {
+                val provider = app.emailPassword
+                provider.registerUser(email, "123456")
+                admin.setCustomConfirmation(true)
+
+                provider.retryCustomConfirmationAsync(email) { result ->
+                    when (result.isSuccess) {
+                        true -> looperThread.testComplete()
+                        false -> fail(result.error.toString())
+                    }
+                }
+            }
+        } finally {
+            admin.setCustomConfirmation(false)
+        }
+    }
+
+    @Test
+    fun retryCustomConfirmation_invalidServerArgsThrows() {
+        val email = "test@10gen.com"
+        admin.setAutomaticConfirmation(false)
+        val provider = app.emailPassword
+        provider.registerUser(email, "123456")
+        admin.setCustomConfirmation(true)
+
+        try {
+            provider.retryCustomConfirmation("foo")
+            fail()
+        } catch (error: AppException) {
+            assertEquals(ErrorCode.USER_NOT_FOUND, error.errorCode)
+        } finally {
+            admin.setCustomConfirmation(false)
+        }
+    }
+
+    @Test
+    fun retryCustomConfirmationAsync_invalidServerArgsThrows() {
+        val email = "test@10gen.com"
+        admin.setAutomaticConfirmation(false)
+        val provider = app.emailPassword
+        provider.registerUser(email, "123456")
+        admin.setCustomConfirmation(true)
+        try {
+            looperThread.runBlocking {
+                provider.retryCustomConfirmationAsync("foo") { result ->
+                    if (result.isSuccess) {
+                        fail()
+                    } else {
+                        assertEquals(ErrorCode.USER_NOT_FOUND, result.error.errorCode)
+                        looperThread.testComplete()
+                    }
+                }
+            }
+        } finally {
+            admin.setCustomConfirmation(false)
+        }
+    }
+
+    @Test
+    fun retryCustomConfirmation_invalidArgumentsThrows() {
+        val provider: EmailPasswordAuth = app.emailPassword
+        assertFailsWith<IllegalArgumentException> { provider.retryCustomConfirmation(TestHelper.getNull()) }
+        looperThread.runBlocking {
+            provider.retryCustomConfirmationAsync(TestHelper.getNull(), checkNullArgCallback)
+        }
+    }
+
+    @Test
     fun sendResetPasswordEmail() {
         val provider = app.emailPassword
         val email: String = "test@10gen.com" // Must be a valid email, otherwise the server will fail
@@ -294,7 +402,7 @@ class EmailPasswordAuthTests {
         provider.registerUser(email, "123456")
         looperThread.runBlocking {
             provider.sendResetPasswordEmailAsync(email) { result ->
-                when(result.isSuccess) {
+                when (result.isSuccess) {
                     true -> looperThread.testComplete()
                     false -> fail(result.error.toString())
                 }
@@ -481,7 +589,7 @@ class EmailPasswordAuthTests {
             provider.resetPasswordAsync("token", TestHelper.getNull(), "password", checkNullArgCallback)
         }
         looperThread.runBlocking {
-            provider.resetPasswordAsync("token","token-id", TestHelper.getNull(), checkNullArgCallback)
+            provider.resetPasswordAsync("token", "token-id", TestHelper.getNull(), checkNullArgCallback)
         }
     }
 
@@ -492,12 +600,13 @@ class EmailPasswordAuthTests {
         val email: String = TestHelper.getRandomEmail()
         for (method in Method.values()) {
             try {
-                when(method) {
+                when (method) {
                     Method.REGISTER_USER -> provider.registerUser(email, "123456")
                     Method.CONFIRM_USER -> provider.confirmUser("token", "tokenId")
                     Method.RESEND_CONFIRMATION_EMAIL -> provider.resendConfirmationEmail(email)
                     Method.SEND_RESET_PASSWORD_EMAIL -> provider.sendResetPasswordEmail(email)
                     Method.CALL_RESET_PASSWORD_FUNCTION -> provider.callResetPasswordFunction(email, "123456")
+                    Method.RETRY_CUSTOM_CONFIRMATION -> provider.retryCustomConfirmation(email)
                     Method.RESET_PASSWORD -> provider.resetPassword("token", "token-id", "password")
                 }
                 fail("$method should have thrown an exception")
@@ -514,12 +623,13 @@ class EmailPasswordAuthTests {
         val callback = App.Callback<Void> { fail() }
         for (method in Method.values()) {
             try {
-                when(method) {
+                when (method) {
                     Method.REGISTER_USER -> provider.registerUserAsync(email, "123456", callback)
                     Method.CONFIRM_USER -> provider.confirmUserAsync("token", "tokenId", callback)
                     Method.RESEND_CONFIRMATION_EMAIL -> provider.resendConfirmationEmailAsync(email, callback)
                     Method.SEND_RESET_PASSWORD_EMAIL -> provider.sendResetPasswordEmailAsync(email, callback)
                     Method.CALL_RESET_PASSWORD_FUNCTION -> provider.callResetPasswordFunctionAsync(email, "123456", arrayOf(), callback)
+                    Method.RETRY_CUSTOM_CONFIRMATION -> provider.retryCustomConfirmationAsync(email, callback)
                     Method.RESET_PASSWORD -> provider.resetPasswordAsync("token", "token-id", "password", callback)
                 }
                 fail("$method should have thrown an exception")